feat(expense_domain): implémente tout les repository de l'app + change le type des methode delete des reposository de Integer à long + ajout de toutes la base d'api pour chaque domaine dans le SecurityFilterChain

+++ davexo-backend/src/main/java/com/davexo/backend/repository/TaskRepository.java
@@ -13,5 +13,5 @@ public interface TaskRepository extends JpaRepository<Task, Integer> {
 
     List<Task> findAllByUserId(Integer userId);
 
-    Integer deleteByIdAndUserId(Integer taskId, Integer userId);
+    long deleteByIdAndUserId(Integer taskId, Integer userId);
 }
